[refactor] rename pkg transaction -> txn
