+++ block/GetCoinbaseParts.go
@@ -41,7 +41,7 @@ import (
 	"encoding/hex"
 	"log"
 
-	"github.com/libsv/libsv/transaction/output"
+	"github.com/libsv/libsv/txn/output"
 	"github.com/libsv/libsv/utils"
 )
