+++ txn/signature/sighash/SigHashFlag.go
